some code clean up
